+++ src/CG.Blazor.Forms._MudBlazor/Attributes/RenderMudRadioGroupAttribute.cs
@@ -8,7 +8,9 @@ using MudBlazor;
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Linq.Expressions;
 using System.Reflection;
+using System.Threading.Tasks;
 
 namespace CG.Blazor.Forms.Attributes
 {
@@ -47,6 +49,18 @@ namespace CG.Blazor.Forms.Attributes
         /// radio buttons in the group.
         /// </summary>
         public string Options { get; set; }
+
+        /// <summary>
+        /// This property contains the name of an optional function, on either the 
+        /// top-level view-model, or the the associated model. That returns a list
+        /// of strings, for options. 
+        /// </summary>
+        /// <remarks>
+        /// <para>
+        /// The method should have this signature: IEnumerable{string} Func()
+        /// </para>
+        /// </remarks>
+        public string OptionsFunc { get; set; }
 
         /// <summary>
         /// This property contains a comma separated list of colors for the 
@@ -98,11 +112,12 @@ namespace CG.Blazor.Forms.Attributes
         public RenderMudRadioGroupAttribute()
         {
             // Set default values.
-            Options = string.Empty;
             Colors = string.Empty;
             Dense = false;
             Disabled = string.Empty;
             DisableRipple = false;
+            Options = string.Empty;
+            OptionsFunc = string.Empty;
             Placement = Placement.End;
             Size = Size.Medium;
         }
@@ -269,12 +284,45 @@ namespace CG.Blazor.Forms.Attributes
                         contentDelegate: childBuilder =>
                         {
                             // Split the optional attributes.
-                            var colors = Colors.Split(',');
-                            var disabled = Disabled.Split(',');
+                            var colors = Colors.Split(',').Select(x => x.Trim()).ToArray();
+                            var disabled = Disabled.Split(',').Select(x => x.Trim()).ToArray();
                             var x = 0;
 
+                            // How should we build the options?
+                            IEnumerable<string> options;
+
+                            // Get the view-model.
+                            var viewModel = path.Last();
+
+                            // Try to resolve the options func.
+                            if (TryOptionsFunc(
+                                viewModel,
+                                propParent,
+                                out var func
+                                ))
+                            {
+                                // If we get here then we resolved the options 
+                                //   func, so let's use it now to populate the
+                                //   options.
+
+                                // Invoke the function.
+                                options = func.Invoke().Result
+                                    .Select(x => x.Trim())
+                                    .ToArray();
+                            }
+                            else
+                            {
+                                // If we get here then we failed to resolve the
+                                //   options func, so, try to use the options
+                                //   property instead.
+
+                                // Split the options.
+                                options = Options.Split(',')
+                                    .Select(x => x.Trim())
+                                    .ToArray();
+                            }
+
                             // Loop through the options.
-                            var options = Options.Split(',');
                             foreach (var option in options)
                             {
                                 var index2 = index; // Reset the index.
@@ -389,6 +437,90 @@ namespace CG.Blazor.Forms.Attributes
                     innerException: ex
                     );
             }
+        }
+
+        #endregion
+
+        // *******************************************************************
+        // Private methods.
+        // *******************************************************************
+
+        #region Private methods
+
+        /// <summary>
+        /// This method checks the <see cref="OptionsFunc"/> property and,
+        /// if the property is populated, searches on the view-model and
+        /// the model for a matching function. That functio is then called
+        /// and the resilts are returned to <paramref name="func"/>.
+        /// </summary>
+        /// <param name="viewModel">The view-model to use for the operation.</param>
+        /// <param name="propParent">The model to use for the operation.</param>
+        /// <param name="func">The output parameter, which is assigned a 
+        /// reference to a func, if one is found.</param>
+        /// <returns>True if the <see cref="OptionsFunc"/> property contains 
+        /// the name of a method on either the view-model, or the model; False
+        /// if not.</returns>
+        private bool TryOptionsFunc(
+            object viewModel,
+            object propParent,
+            out Func<Task<IEnumerable<string>>> func
+            )
+        {
+            func = null;
+
+            // Is the property populated?
+            if (false == string.IsNullOrEmpty(OptionsFunc))
+            {
+                // Create possible targets for the search.
+                var targets = (viewModel == propParent)
+                    ? new[] { viewModel }
+                    : new[] { viewModel, propParent };
+
+                // Loop and look for the function.
+                foreach (var target in targets)
+                {
+                    // Get the target type.
+                    var targetType = target.GetType();
+
+                    // Look for the named method.
+                    var methodInfo = targetType.GetMethod(
+                        OptionsFunc,
+                        BindingFlags.Public |
+                        BindingFlags.NonPublic |
+                        BindingFlags.Instance
+                        );
+
+                    // Did we succeed?
+                    if (null != methodInfo)
+                    {
+                        // Create a viewModel reference expression.
+                        var viewModelExp = Expression.Constant(
+                            target
+                            );
+
+                        // Create the method call expression.
+                        var callExp = Expression.Call(
+                            viewModelExp,
+                            methodInfo
+                            );
+
+                        // Create a lambda expression.
+                        var lambdaExp = Expression.Lambda<Func<Task<IEnumerable<string>>>>(
+                            callExp,
+                            callExp.Arguments.OfType<ParameterExpression>()
+                            );
+
+                        // Compile the expression to a func.
+                        func = lambdaExp.Compile();
+
+                        // We found the func.
+                        return true;
+                    }
+                }
+            }
+
+            // We didn't find the func.
+            return false;
         }
 
         #endregion