+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 
 * bug fixes
 
+* I added an OptionsExp property to the RenderMudSelectAttribute and RenderMudRadioGroupAttribute classes to allow fetching options from a live data source.
+